Milestone 1 (part 2) (#1)
* Formatting and some code cleanup

* Fix css

* Add form validation

* Removed some code

* Setup jasmine spec

Add links to jasmine libraries and CSS by creating a test HTML file

* Add tests

Added tests and a helper function that will add HTML elements necesary for the tests. A test DIV element is added before each test and the test elements are mounted in it. After each test, this DIV is destroyed and thus the test elements are removed with it.

* Separate functions

The click handler code has been moved to separate functions to make it easier to test click events by invoking the handler functions directly.

* Add more tests

* Removed unused code

* Hide the currency conversion section after 5 seconds

* Remove redundant code

* Added different sections to README

* Readme updated